% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SBMLtoOdin.R
\name{AddToParamLib}
\alias{AddToParamLib}
\title{Title}
\usage{
AddToParamLib(r, local_param_lib, reserved_lib)
}
\arguments{
\item{r}{An integer.}

\item{local_param_lib}{A dictionary.}

\item{reserved_lib}{A dictionary.}
}
\value{
A dictionary.
}
\description{
Title
}
