% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SBMLtoOdin.R
\name{getFunctionOutput}
\alias{getFunctionOutput}
\title{Title}
\usage{
getFunctionOutput(m, ind, r)
}
\arguments{
\item{m}{A libSBML model object.}

\item{ind}{An integer.}

\item{r}{An integer.}
}
\description{
Title
}
