% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SBMLtoOdin.R
\name{getFunctionOutputForRules}
\alias{getFunctionOutputForRules}
\title{Title}
\usage{
getFunctionOutputForRules(m, formula, func_id)
}
\arguments{
\item{m}{A libSBML model object.}

\item{ind}{A string.}

\item{r}{A string.}
}
\value{
function output of rule as string
}
\description{
Title
}
\examples{
getFunctionOutputForRules(model, formula, index)
}
\keyword{internal}
