% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SBMLtoOdin.R
\name{getFunctionParams}
\alias{getFunctionParams}
\title{Title}
\usage{
getFunctionParams(file_content, r, param_lib, reserved_lib)
}
\arguments{
\item{file_content}{A string (file content of odin model file).}

\item{r}{An integer.}

\item{param_lib}{A dictionary.}

\item{reserved_lib}{A dictionary.}
}
\value{
A string.
}
\description{
Title
}
\examples{
getFunctionParams(file_content, reaction, param_dictionary, reserved_names_dictionary)
}
