% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SBMLtoOdin.R
\name{getSpeciesRule}
\alias{getSpeciesRule}
\title{Title}
\usage{
getSpeciesRule(m, i, species_dic, func_def_dict)
}
\arguments{
\item{m}{A libSBML model object.}

\item{i}{An integer for the index of the rule.}

\item{species_dic}{The dictionary containing all species.}

\item{func_def_dict}{The dictionary containing all function definitions.}
}
\value{
species dictionary.
}
\description{
Title
}
\examples{
getSpeciesRule(model,1,dictionary)
}
