% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SBMLtoOdin.R
\name{importSBML}
\alias{importSBML}
\title{Title}
\usage{
importSBML(path_to_input)
}
\arguments{
\item{path_to_input}{A string (Path to an SBML model file).}
}
\value{
A libSBML model object.
}
\description{
Title
}
\examples{
importSBML("/usr/model_files/my_model.xml")
}
