% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SBMLtoOdin.R
\name{importSBMLfromBioModels}
\alias{importSBMLfromBioModels}
\title{Title}
\usage{
importSBMLfromBioModels(model_id, path_to_output = "odinModel.R")
}
\arguments{
\item{model_id}{A string (a valid BioModels id)}

\item{path_to_output}{A string (Path to the output file).}
}
\description{
Title
}
