% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SBMLtoOdin.R
\name{importSBMLfromFile}
\alias{importSBMLfromFile}
\title{Title}
\usage{
importSBMLfromFile(path_to_input, path_to_output = "odinModel.R")
}
\arguments{
\item{path_to_input}{A string (Path to an SBML model file).}

\item{path_to_output}{A string (Path to the output file).}
}
\value{
no return
}
\description{
Title
}
\examples{
importSBMLfromFile("/usr/model_files/my_model.xml", "/usr/model_files/my_model.R")
}
