% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SBMLtoOdin.R
\name{in_reserved_lib}
\alias{in_reserved_lib}
\title{Title}
\usage{
in_reserved_lib(param_name, reserved_lib)
}
\arguments{
\item{param_name}{A string.}

\item{reserved_lib}{A dictionary.}
}
\value{
A string.
}
\description{
Title
}
\examples{
in_reserved_lib("a", c("a"="a1", "b"="b1"))
}
\keyword{internal}
