% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SBMLtoOdin.R
\name{SBML_to_odin}
\alias{SBML_to_odin}
\title{Title}
\usage{
SBML_to_odin(path_to_input, path_to_output = "odinModel.R")
}
\arguments{
\item{path_to_input}{A string (location of SBML model file).}

\item{path_to_output}{A string (location of odin model file to be created).}
}
\description{
Title
}
\examples{
SBMLtoOdin("/usr/models/my_SBML_model.xml","/usr/models/my_odin_model.R")
}
