% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SBMLtoOdin.R
\name{sub_and}
\alias{sub_and}
\title{Title}
\usage{
sub_and(file_content)
}
\arguments{
\item{file_content}{A string.}
}
\value{
file content
}
\description{
Title
}
\examples{
sub_and("if(and(b,c))")
}
\keyword{internal}
