% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SBMLtoOdin.R
\name{sub_leq}
\alias{sub_leq}
\title{Title}
\usage{
sub_leq(file_content)
}
\arguments{
\item{file_content}{A string.}
}
\value{
file content
}
\description{
Title
}
\examples{
sub_leq("if(leq(a,2))")
}
\keyword{internal}
