% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SBMLtoOdin.R
\name{sub_lt}
\alias{sub_lt}
\title{Title}
\usage{
sub_lt(file_content)
}
\arguments{
\item{file_content}{A string.}
}
\value{
file content A string.
}
\description{
Title
}
\examples{
sub_lt("if(lt(a,2))")
}
\keyword{internal}
