% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SBMLtoOdin.R
\name{translate_piecewise}
\alias{translate_piecewise}
\title{Title}
\usage{
translate_piecewise(file_content)
}
\arguments{
\item{file_content}{A string.}
}
\value{
file content as string
}
\description{
Title
}
\examples{
translate_piecewise("piecewise(1.55, lt(C, 3100), 0.2)")
}
\keyword{internal}
