% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SBMLtoOdin.R
\name{translate_root}
\alias{translate_root}
\title{Title}
\usage{
translate_root(file_content)
}
\arguments{
\item{file_content}{A string.}
}
\value{
file content as string
}
\description{
Title
}
\examples{
translate_root("root(2,4)")
}
