% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SBMLtoOdin.R
\name{getRule}
\alias{getRule}
\title{Title}
\usage{
getRule(
  file_content,
  m,
  i,
  var_params_dict_used,
  var_params_dict,
  reserved_names_lib
)
}
\arguments{
\item{file_content}{A string.}

\item{m}{A libSBML model object.}

\item{i}{An integer.}

\item{var_params_dict_used}{dictionary that saves param usage}

\item{var_params_dict}{dictionary that has parameter (initial) values}
}
\value{
A string (updated file odin file content)
}
\description{
Title
}
\examples{
getRule("initial(S1) <- S1_init", model, 1)
}
