% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SBMLtoOdin.R
\name{getSpeciesRule}
\alias{getSpeciesRule}
\title{Title}
\usage{
getSpeciesRule(m, i, species_dic, func_def_dict)
}
\arguments{
\item{func_def_dict}{}
}
\description{
Title
}
\examples{
getSpeciesRule(model,1,dictionary)
}
