% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SBMLtoOdin.R
\name{SBML_to_odin}
\alias{SBML_to_odin}
\title{Title}
\usage{
SBML_to_odin(model, path_to_output, input_str)
}
\arguments{
\item{model}{A libSBML model.}

\item{path_to_output}{A string (location of odin model file to be created).}
}
\value{
no return
}
\description{
Title
}
\examples{
SBMLtoOdin("/usr/models/my_SBML_model.xml","/usr/models/my_odin_model.R")
}
\keyword{internal}
